refactor(cli): move schedule catalog subtree to schedule_catalog.go (PRD #1009 M2)
